+++ relay-server/src/actors/outcome_aggregator.rs
@@ -187,8 +187,12 @@ impl Service for OutcomeAggregator {
 
             loop {
                 tokio::select! {
-                    Some(message) = rx.recv() => self.handle_track_outcome(message),
+                    // Prioritize flush over receiving messages to prevent starving. Shutdown can be
+                    // last since it is not vital if there are still messages in the channel.
+                    biased;
+
                     () = &mut self.flush_handle => self.flush(),
+                    Some(message) = rx.recv() => self.handle_track_outcome(message),
                     _ = shutdown.changed() => self.handle_shutdown(&shutdown.borrow_and_update()),
                     else => break,
                 }