+++ relay-server/src/utils/actix.rs
@@ -15,28 +15,12 @@ impl<T, E> Response<T, E> {
         Response::Reply(Ok(value))
     }
 
-    pub fn reply(result: Result<T, E>) -> Self {
-        Response::Reply(result)
-    }
-
     pub fn future<F>(future: F) -> Self
     where
         F: IntoFuture<Item = T, Error = E>,
         F::Future: 'static,
     {
         Response::Future(Box::new(future.into_future()))
-    }
-}
-
-impl<T: 'static, E: 'static> Response<T, E> {
-    pub fn map<U, F: 'static>(self, f: F) -> Response<U, E>
-    where
-        F: FnOnce(T) -> U,
-    {
-        match self {
-            Response::Reply(result) => Response::reply(result.map(f)),
-            Response::Future(future) => Response::future(future.map(f)),
-        }
     }
 }
 
@@ -65,19 +49,20 @@ where
     }
 }
 
-/// Constructs a single threaded tokio [`Runtime`] containing a clone of the actix [`System`].
+/// Constructs a tokio [`Runtime`] configured for running [services](relay_system::Service).
 ///
-/// This is required if you need to send messages from the tokio runtime to actix
-/// actors.
+/// For compatibility, this runtime also registers the actix [`System`]. This is required for
+/// interoperability with actors. To send messages to those actors, use
+/// [`compat::send`](relay_system::compat::send).
 ///
 /// # Panics
 ///
-/// The calling thread must have the actix system enabled, panics if this is invoked
-/// in a thread where actix is not enabled.
-pub fn tokio_runtime_with_actix() -> Runtime {
+/// The calling thread must have the actix system enabled, panics if this is invoked in a thread
+/// where actix is not enabled.
+pub fn create_runtime(threads: usize) -> Runtime {
     let system = System::current();
     tokio::runtime::Builder::new_multi_thread()
-        .worker_threads(1)
+        .worker_threads(threads)
         .enable_all()
         .on_thread_start(clone!(system, || System::set_current(system.clone())))
         .build()